SPT: Move Template Selector files into separate directory
