+++ apps/full-site-editing/full-site-editing-plugin/starter-page-templates/page-template-modal/components/template-selector/block-preview.js
